Add AssessmentAdapter
